### 2017-07-14 #319 @whedon
[PRE REVIEW]: geneXplainR: An R interface for the geneXplain platform
**Submitting author:** @pstegma (<a href="http://orcid.org/0000-0002-1881-8746">Philip Stegmaier</a>)
**Repository:** <a href="https://github.com/genexplain/geneXplainR" target ="_blank">https://github.com/genexplain/geneXplainR</a>
**Version:** 0.4
