### 2017-07-24 #319 @mgymrek
@pstegma thanks for the submission. I have taken a look at the paper.md. Could you clarify what types of analyses can be done with this platform? The paper currently mentions "a broad range of bioinformatic and systems biology applications" and "upstream anslysis", but it is not clear exactly what types of analyses are being targeted. e.g. analysis of sequencing data? gene ontology analysis? Some examples in the paper will be helpful to readers to determine whether the tool is relevant to them, and will also help me find an appropriate reviewer.
