### 2017-08-07 350017a
Add information about methods and references
