+++ paper/paper.bib
@@ -11,7 +11,8 @@
   title={“Upstream Analysis”: An integrated promoter-pathway analysis approach to causal interpretation of microarray data},
   author={Koschmann, J. and Bhar, A. and Stegmaier,P. and Kel, A. E. and Wingender, E.},
   journal={Microarrays},
-  number={4},
+  number={2},
+  volumne={4},
   pages={270-286},
   year={2015}
 }
@@ -45,9 +46,68 @@
   title={Bioconductor: open software development for computational biology and bioinformatics},
   author={Gentleman, R.C. and Carey, V.J. and Bates, D.M. and Bolstadm B. and Dettling, M. and others},
   journal={Genome Biology},
-  number={5},
-  volume={10},
+  number={10},
+  volume={5},
   pages={R80},
   year={2004}
+}
+
+@article{GO,
+    title={Gene ontology: tool for the unification of biology. The Gene Ontology Consortium},
+    author={Ashburner, M. and Ball, C.A. and Blake, J.A. and Botstein, D. and Butler, H.},
+    journal={Nature Genetics},
+    number={1},
+    volume={25},
+    pages={25-29}
+}
+
+@article{GSEA,
+    title={Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles},
+    author={Subramanian, A. and Tamayo, P. and Mootha, V.K. and Mukherjee, S. and Ebert, B.L. and others},
+    journal={Proc Natl Acad Sci U S A},
+    number={43},
+    volume={102},
+    pages={15545-15550},
+    year={2005}
+}
+
+@article{FBA,
+    title={Global reconstruction of the human metabolic network based on genomic and bibliomic data},
+    author={Duarte, N. and Becker, S.A. and Jamshidi, N. and Thiele, I. and Mo, M.L. and others},
+    journal={Proc Natl Acad Sci U S A},
+    number={6},
+    volume={104},
+    pages={1777-1782},
+    year={2007}
+}
+
+@article{limma,
+    title={limma powers differential expression analyses for RNA-sequencing and microarray studies},
+    author={Ritchie, M.E. and Phipson, B. and  Wu, D. and Hu, Y. and Law, C.W. and others},
+    journal={Nucleic Acids Research},
+    number={7},
+    volume={43},
+    pages={e47},
+    year={2015}
+}
+
+@article{deseq,
+    title={Differential expression analysis for sequence count data},
+    author={Anders, S. and Huber, W.},
+    journal={Genome Biology},
+    number={10},
+    volume={11},
+    pages={R106},
+    year={2010}
+}
+
+@article{biomodels,
+    title={BioModels Database: A Free, Centralized Database of Curated, Published, Quantitative Kinetic Models of Biochemical and Cellular Systems},
+    author={Le Novère, N. and Bornstein, B. and Broicher, A. and Courtot, M. and  Donizelli, M. and others},
+    journal={Nucleic Acids Research},
+    number={Database issue},
+    volume={34},
+    pages={D689-691},
+    year={2006}
 }
 
