### 2017-08-07 #319 @pstegma
Dear @mgymrek ,
thank you for taking a look at our paper and for your suggestions. I have added information about covered methods. As we are part of a few consortia, can I be of any help to find a reviewer?
