### 2017-09-08 #319 @gymreklab
@iugrina @selbouhaddani - I believe you have the background to review this submission by @pstegma (https://github.com/genexplain/geneXplainR). Are you interested in reviewing this submission for JOSS? For more background on the review process, see: http://joss.theoj.org/about#reviewer_guidelines
