### 2017-09-27 #412 @selbouhaddani
# General

The submission describes an R package "*geneXplainR*" that provides an interface to connect to an external platform "geneXplain". By using the R package queries can be submitted and results can be retrieved from the platform. This can be done without leaving the R environment.

Overall I think this is a submission suited for the JOSS. The R package enables access to bioinformatics tools and workflows used by bioinformatics experts and researchers from R, from which additional R functionalities can be used to further process the results. Although the platform is closed source, the R package is open source and there is an option to freely register to get an account.

# Comments

## Usage

- There is a small typo in the README.md: `gx.login("http://genexample.com/bioumlweb","user","password"),`
Please remove the last comma.
- The first argument of the function `gx.login` in the README.md is `"http://genexample.com/bioumlweb"` which does not work (I got an error: `Could not resolve host: genexample.com`). Perhaps an alternative is to just use `"https://platform.genexplain.com"` as this is the default value. In addition I couldn't find any other web addresses in the documentation that could also be inserted as argument, it is a good idea to give examples e.g. in the help file of `gx.login`.
- In section **Example** in the README.md in the *examples* branch a code chunk is shown in which a functional classification analysis of a gene set is discussed. Currently this chunk only works if the user already has uploaded some dataset in the correct format.
It's not strictly necessary, but I think availability of a toy dataset would be very appreciated by the user, so that this code runs without having to construct and upload datasets. This would help understanding the code and the functionalities.

## Paper.md

In my opinion the paper.md file is well written and well structured. Some minor comments are:

- In "*Methods integrated into the system integrates include*", the word "*integrates*" should be removed.
- Unfortunately it is not clear where to report issues or seek support. Should the user use the Github Issue system, or go to the genexplain website? Please add a statement to clarify this.
- The citations in the `paper.bib` file do not include doi's, while they are available (e.g. the article corresponding to *@upstreamAnalysis* has doi 10.3390/microarrays4020270). If I interpret the last checkmark correctly, the available doi's should be added.

Please let me know if anything was unclear.
