### 2017-10-05 #412 @pstegma
Thank you for the comments. I am working on the suggested improvements.
