### 2017-10-05 dfa5bf9
Fix typo in README.md. Correct platform.genexplain.com.
