### 2017-10-05 d972f97
Add instructions for gx.login without account
