### 2017-10-05 b1708c9
Add examples to gx.login
