### 2017-10-05 6f702b1
Update Example section of README.md to give working example
