+++ README.md
@@ -66,22 +66,25 @@ Performing a functional classification analysis on a gene set may
 look like this:
 
 ```R
-## Let us have data located in a project named My_Project.
-my_data <- "data/Projects/My_Project/Data/First_data"
+## Login to demo account
+gx.login()
 
-## This will be the path to the result table
-my_result <- paste(my_data,"_fc_result",sep="")
+## Example data sets can be found under 'data/Examples'. Here we use a set of upregulated genes.
+my_data <- "data/Examples/TNF-stimulation of HUVECs GSE2639, Affymetrix HG-U133A microarray/Data/DEGs with limma/Normalized (RMA) DEGs with limma/Condition_1 vs. Condition_2/Up-regulated genes Ensembl"
+
+## This will be the path to the result table. It will be put into the 'Demo project' provided in this workspace.
+my_result <- paste0("data/Projects/Demo project/Data/","my_demo_result_",Sys.Date(),"_",floor(runif(1, 1, 10^12)))
 
 ## This invokes the functional classification analysis according
-## to disease/gene assocations in HumanPSD.
+## to the Gene Ontology.
 gx.analysis("Functional classification",
              list(sourcePath = my_data,
              species = "Human (Homo sapiens)",
-             bioHub = "HumanPSD(TM) disease",
+             bioHub = "Full gene ontology classification",
              outputTable = my_result));
                 
 ## This command loads the result table into an R data frame.
 data <- gx.get(my_result)
 ```
 
-  
+   
