### 2017-10-05 90daa9c
Add Support section with indication where to report issues.
