+++ paper/paper.md
@@ -78,4 +78,9 @@ by the European Union Seventh Framework Programme (FP7/2007-2013) under grant
 agreement no. 305280, research area FP7-HEALTH-2012-INNOVATION-1, topic 
 HEALTH.2012.2.1.1-3: Statistical methods for collection and analysis of –omics data.
 
+## Support
+
+If you find an issue, we would be happy if you let us know either by writing an e-mail
+to <info@genexplain.com> or to the GitHub issue system (<https://github.com/genexplain/geneXplainR/issues>).
+
 # References
