### 2017-10-05 29436f3
Add Support section with indication where to report issues to README.md
