### 2017-10-05 5db7c1d
Update Support section in README.md
