### 2017-10-05 c74475d
Add dois and doi URLs
