+++ paper/paper.bib
@@ -4,7 +4,7 @@
   journal={J. Biomol. Tech.},
   number={22},
   pages={S1},
-  year={2011}
+  year={2011},
 }
 
 @article{upstreamAnalysis,
@@ -14,7 +14,9 @@
   number={2},
   volumne={4},
   pages={270-286},
-  year={2015}
+  year={2015},
+  doi={10.3390/microarrays4020270},
+  url={http://doi.org/10.3390/microarrays4020270}
 }
 
 @manual{R,
@@ -24,6 +26,7 @@
   address = {Vienna, Austria},
   year = {2016},
   url = {https://www.R-project.org/},
+
 }
 
 @manual{rbiouml,
@@ -49,7 +52,9 @@
   number={10},
   volume={5},
   pages={R80},
-  year={2004}
+  year={2004},
+  doi={10.1186/gb-2004-5-10-r80},
+  url={https://doi.org/10.1186/gb-2004-5-10-r80}
 }
 
 @article{GO,
@@ -58,7 +63,9 @@
     journal={Nature Genetics},
     number={1},
     volume={25},
-    pages={25-29}
+    pages={25-29},
+    doi={10.1038/75556},
+    url={https://doi.org/10.1038/75556}
 }
 
 @article{GSEA,
@@ -68,7 +75,9 @@
     number={43},
     volume={102},
     pages={15545-15550},
-    year={2005}
+    year={2005},
+    doi={10.1073/pnas.0506580102},
+    url={https://doi.org/10.1073/pnas.0506580102}
 }
 
 @article{FBA,
@@ -78,7 +87,9 @@
     number={6},
     volume={104},
     pages={1777-1782},
-    year={2007}
+    year={2007},
+    doi={10.1073/pnas.0610772104},
+    url={https://doi.org/10.1073/pnas.0610772104}
 }
 
 @article{limma,
@@ -88,7 +99,9 @@
     number={7},
     volume={43},
     pages={e47},
-    year={2015}
+    year={2015},
+    doi={10.1093/nar/gkv007},
+    url={https://doi.org/10.1093/nar/gkv007}
 }
 
 @article{deseq,
@@ -98,7 +111,9 @@
     number={10},
     volume={11},
     pages={R106},
-    year={2010}
+    year={2010},
+    doi={10.1186/gb-2010-11-10-r106},
+    url={https://doi.org/10.1186/gb-2010-11-10-r106}
 }
 
 @article{biomodels,
@@ -108,6 +123,8 @@
     number={Database issue},
     volume={34},
     pages={D689-691},
-    year={2006}
+    year={2006},
+    doi={10.1093/nar/gkj092},
+    url={https://doi.org/10.1093/nar/gkj092}
 }
 
