### 2017-10-05 #412 @pstegma
First of all, many thanks again to the reviewer for spending time with the review and for his very helpful comments which we are certain help to improve the quality of the software. We think we were able to implement all of the requested points. Please find below our detailed response.

# Replies to reviewer comments

## Usage

The comma typo as well as the misspelled URL were corrected. More examples
and instructions how to call the `gx.login` function were added to the README.md as well as to the geneXplain.R source file.
The following sentences were added to the Usage section of the README.md:

```Markdown
If you just wish to try out the functionality of geneXplainR, you can sign into the demo account on our
public server *platform.genexplain.com* by simply calling *gx.login()*, without parameters. To access another server that
provides a demo workspace, you can call *gx.login* with the server URL as only argument.
```

The following executable examples were added to the roxygen of `gx.login` in R/geneXplain.R:

```R
 #'
 #' @examples
 #'
 #' # Sign into demo account on public server
 #' gx.login()
 #' gx.logout()
 #' # Sign into demo account specifying server URL
 #' gx.login("https://platform.genexplain.com")
 #' gx.logout()
 #'
```

The Example section in the README.md of the *examples* branch was updated
and now contains the following executable R code which entirely relies on
existing example data and demo workspace:

```R
## Login to demo account
gx.login()

## Example data sets can be found under 'data/Examples'. Here we use a set of upregulated genes.
my_data <- paste0("data/Examples/TNF-stimulation of HUVECs GSE2639, ",
                  "Affymetrix HG-U133A microarray/Data/DEGs with limma/",
                  "Normalized (RMA) DEGs with limma/Condition_1 vs. Condition_2/",
                  "Up-regulated genes Ensembl")

## This will be the path to the result table. It will be put into the 'Demo project' provided
##  in this workspace.
my_result <- paste0("data/Projects/Demo project/Data/",
                    "my_demo_result_",
                    Sys.Date(),"_",
                    floor(runif(1, 1, 10^12)))

## This invokes the functional classification analysis according
## to the Gene Ontology.
gx.analysis("Functional classification",
             list(sourcePath = my_data,
             species = "Human (Homo sapiens)",
             bioHub = "Full gene ontology classification",
             outputTable = my_result));

## This command loads the result table into an R data frame.
data <- gx.get(my_result)
```

## Paper.md

Support sections describing where to report issues were added to the README.md of the master branch as well as to the paper.

The dois and doi URLs were added to citations in paper.bib.
