### 2017-10-09 1671c7d
Update geneXplain.R
