### 2017-10-09 a1d0bb7
Temporarily omit gx.isElement test
