### 2017-10-09 9d76bd2
Add gx.isElement test
