+++ tests/testthat/test_geneXplainR.R
@@ -399,9 +399,17 @@ test_that("gx.enrichedUpstreamAnalysis", {
         })
 
 test_that("gx.exists",{
-  expect_output(gx.exists("gx.put"), "TRUE")
-  expect_output(gx.exists("put"),"TRUE")
-  expect_output(gx.exists("42"),"FALSE")
+              expect_output(gx.exists("gx.put"), "TRUE")
+              expect_output(gx.exists("put"),"TRUE")
+              expect_output(gx.exists("42"),"FALSE")
 })
 
-
+test_that("gx.isElement",{
+              before.test.connection()
+              bioumlQueryValue <<- "{\"type\":0,\"values\":\"{\\\"names\\\":[{\\\"name\\\":\\\"Data\\\"},{\\\"name\\\":\\\"More\\\"}]}\"}"
+              expect_true(gx.isElement("test.folder", "Data"))
+              expect_false(gx.isElement("test.folder", "42"))
+              expect_true(gx.isElement("test.folder", "More"))
+              expect_equal(as.character(queryJsonArgs$params['dc']), "test.folder")
+              clean.after.test()
+})
