### 2017-10-09 #412 @selbouhaddani
Thank you for addressing all the remarks. I went through the package again and verified that all JOSS requirements are met.
@mgymrek I'm satisfied with the current submission.
